gcloudrun/requirements.txt
functions-framework==3.*
google-cloud-firestore
flask

google-generativeai
python-dotenv
